requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
pandas~=2.2.1
PyJWT==2.8.0
pyodbc==5.1.0
pytz==2024.1
requests==2.31.0
six==1.16.0
yarl==1.9.4

matplotlib~=3.8.3
selenium~=4.18.1
beautifulsoup4~=4.12.3
icmplib~=3.0.4
code128~=0.3
mlxtend~=0.23.1

aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asttokens==2.4.1
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
code128==0.3
colorama==0.4.6
cors==1.0.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
docopt==0.6.2
docx==0.2.4
docxcompose==1.4.0
docxtpl==0.16.8
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Limiter==3.6.0
frozenlist==1.4.1
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.6
importlib_resources==6.4.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
limits==3.11.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
ordered-set==4.1.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pika==1.3.2
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyodbc==5.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
pywin32==306
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
rpds-py==0.18.1
setuptools==69.2.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.3.0
tldextract==5.1.2
tornado==6.4.1
traitlets==5.14.3
twilio==9.0.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
zope.event==5.0
zope.interface==6.2